ref(dashboard): Refactor Dashboard details
